feat: use integer data type for transaction amount instead of money

+++ src/modules/transaction/entities/transaction.entity.ts
@@ -12,7 +12,7 @@ export class Transaction {
   @ManyToOne(() => User)
   user: UserDto;
 
-  @Column('money')
+  @Column('integer')
   amount: number;
 
   @Column({
